moved gui includes
