+++ executable/gui/em_fitter_gui/em_fitter_gui.cpp
@@ -17,9 +17,9 @@
 #include <shell/Command.h>
 #include <settings/All.h>
 
-#include <logo.h>
-#include <include.h>
-#include <resources.h>
+#include <gui/helper.h>
+#include <gui/logo.h>
+#include <gui/resources.h>
 
 #include <filesystem>
 #include <algorithm>
