+++ executable/gui/saxs_fitter_gui/CMakeLists.txt
@@ -8,5 +8,6 @@ set(ELEMENTS_APP_VERSION "1.0")
 set(ELEMENTS_APP_SOURCES ${sources} ${CMAKE_CURRENT_SOURCE_DIR}/saxs_fitter_gui.cpp)
 include(ElementsConfigApp)
 
-target_link_libraries("${ELEMENTS_APP_PROJECT}" PRIVATE ausaxs_core ausaxs_math)
+target_link_libraries("${ELEMENTS_APP_PROJECT}" PRIVATE ausaxs_core ausaxs_math nfd)
+target_include_directories("${ELEMENTS_APP_PROJECT}" PRIVATE "${CMAKE_SOURCE_DIR}/include/gui")
 add_dependencies("${ELEMENTS_APP_PROJECT}" plotter)
