+++ include/gui/gui/resources.h
